+++ src/spox/_standard.py
@@ -1,5 +1,4 @@
 """Module implementing a base for standard ONNX operators, which use the functionality of ONNX node-level inference."""
-import logging
 from typing import TYPE_CHECKING, Callable, Dict, Tuple
 
 import numpy
@@ -175,15 +174,9 @@ class StandardNode(Node):
             for var in self.inputs.get_vars().values()
             if var._value
         }
-        try:
-            output_feed = run(model, input_feed)
-        except Exception as e:
-            logging.debug(
-                f"Value propagation in {self.get_op_repr()} on backend "
-                f"{_value_prop._VALUE_PROP_BACKEND} failed with - "
-                f"{type(e).__name__}: {e}"
-            )
-            output_feed = {}
+
+        output_feed = run(model, input_feed)
+
         results = {
             scope.var[str(name)]
             ._which_output: unwrap_feed(scope.var[str(name)].unwrap_type(), result)